refactor(search page)
